Use ref instead of findDOMNode in jazzicon component
The jazzicon component was using `findDOMNode` to get the DOM node for
the main div returned by the component, which is generally not
recommended. Instead a ref is now used.

This also means that the component works correctly under test, so the
`detect-node` package is no longer required to skip that behavior when
running unit tests.